fix: fix devsettings
