+++ packages/kit/src/components/PerformanceMonitor/MonitorContainer.native.tsx
@@ -1,0 +1,3 @@
+export const MonitorContainer = () => {
+  return null;
+};
